+++ src/api/hooks/useGetProducts.test.tsx
@@ -2,7 +2,7 @@ import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { renderHook, waitFor } from '@testing-library/react';
 import React from 'react';
 
-import { PRODUCTS_MOCK_DATA } from './products.mock';
+import { PRODUCTS_MOCK_DATA } from '../mocks/products.mock';
 import { useGetProducts } from './useGetProducts';
 
 describe('useGetProducts', () => {
