+++ src/api/mocks/productDetail.mock.ts
