+++ src/mocks/browser.ts
@@ -1,12 +1,12 @@
 import { setupWorker } from 'msw';
 
-import { categoriesMockHandler } from '@/api/hooks/categories.mock';
-import { loginMockHandler } from '@/api/hooks/login.mock';
-import { productDetailMockHandler } from '@/api/hooks/productDetail.mock';
-import { productOptionsMockHandler } from '@/api/hooks/productOptions.mock';
-import { productsMockHandler } from '@/api/hooks/products.mock';
-import { registerMockHandler } from '@/api/hooks/register.mock';
-import { wishlistMockHandlers } from '@/api/hooks/wishlist.mock';
+import { categoriesMockHandler } from '@/api/mocks/categories.mock';
+import { loginMockHandler } from '@/api/mocks/login.mock';
+import { productDetailMockHandler } from '@/api/mocks/productDetail.mock';
+import { productOptionsMockHandler } from '@/api/mocks/productOptions.mock';
+import { productsMockHandler } from '@/api/mocks/products.mock';
+import { registerMockHandler } from '@/api/mocks/register.mock';
+import { wishlistMockHandlers } from '@/api/mocks/wishlist.mock';
 export const worker = setupWorker(
   ...categoriesMockHandler,
   ...productsMockHandler,
