+++ src/pages/Register/index.tsx
@@ -2,7 +2,7 @@ import styled from '@emotion/styled';
 import { useState } from 'react';
 import { Link, useNavigate } from 'react-router-dom';
 
-import { useGetRegister } from '@/api/hooks/useRegister';
+import { useGetRegister } from '@/api/hooks/useGetRegister';
 import KAKAO_LOGO from '@/assets/kakao_logo.svg';
 import { Button } from '@/components/common/Button';
 import { UnderlineTextField } from '@/components/common/Form/Input/UnderlineTextField';
